mv cli pkg to cmd, reducing package #
